+++ src/main/java/dev/willyelton/crystal_tools/item/armor/LevelableArmor.java
@@ -143,6 +143,16 @@ public class LevelableArmor extends ArmorItem implements LevelableItem, Wearable
     }
 
     @Override
+    public boolean isEnchantable(@NotNull ItemStack stack) {
+        return false;
+    }
+
+    @Override
+    public boolean isBookEnchantable(ItemStack stack, ItemStack book) {
+        return false;
+    }
+
+    @Override
     public void onArmorTick(ItemStack stack, Level world, Player player) {
         if (this.isDisabled()) {
             stack.shrink(1);