+++ changelog.md
@@ -3,10 +3,12 @@
 - Updates to Forge 40.2.17
 
 ### Additions
+- Adds config option `REPAIR_IN_HAND`: If true, tools can repair while you are holding them
 
 ### Fixes
 - Fixes recipe for crystal elytra and crystal AIOT not showing up in JEI
 - Fixes the crystal bow not being able to shoot arrows from quivers
+- Fixes tools and armor being enchantable
 
 ----------------------- 1.3.0 Changes -----------------------
 ### Additions